small enhancements

+++ src/Feature/RS485/RBUS/Data.hs
@@ -32,7 +32,7 @@ data RBUS = forall t. LazyTransport t => RBUS
      , msgBuff          :: Buffer 512 Uint16
      , msgIndex         :: Value      Uint16
      , txBuff           :: Buffer 255 Uint16
-     , rsBuff           :: Buffer 255 Uint16
+     , rsBuff           :: Buffer 255 Uint8
      , rsSize           :: Value      Uint8
      , rxLock           :: Value      IBool
      , txLock           :: Value      IBool
